requirements.txt
accelerate==0.31.0
addict==2.4.0
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
crcmod==1.7
cryptography==42.0.8
cycler==0.12.1
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.14.0
einops==0.8.0
filelock==3.14.0
fonttools==4.53.0
fsspec==2024.6.0
ftfy==6.2.0
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==8.0.0
intel-openmp==2021.4.0
Jinja2==3.1.4
jmespath==0.10.0
kiwisolver==1.4.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mat4py==0.6.0
matplotlib==3.9.0
mdurl==0.1.2
mkl==2021.4.0
mmcv==2.0.1
mmdet==3.0.0
mmengine==0.10.3
mmpretrain==1.2.0
mmsegmentation==1.2.2
model-index==0.1.11
modelindex==0.0.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.3
-e git+https://github.com/likyoo/open-cd.git@d66810589fb1269f1a1eef0c0ffb706480c6df47#egg=opencd
opencv-python==4.10.0.84
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.0
ordered-set==4.1.0
oss2==2.17.0
packaging==24.1
pandas==2.2.2
peft==0.11.1
pillow==10.3.0
platformdirs==4.2.2
prettytable==3.10.0
psutil==6.0.0
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2023.4
pywin32==306
PyYAML==6.0.1
regex==2024.5.15
requests==2.28.2
rich==13.4.2
safetensors==0.4.3
scipy==1.14.0
shapely==2.0.4
six==1.16.0
sqlparse==0.5.0
sympy==1.12.1
tabulate==0.9.0
tbb==2021.13.0
termcolor==2.4.0
terminaltables==3.1.10
tokenizers==0.19.1
tomli==2.0.1
torch==2.0.0
torchaudio==2.0.1+cpu
torchvision==0.15.1+cpu
tqdm==4.65.2
transformers==4.41.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
wcwidth==0.2.13
yapf==0.40.2
zipp==3.19.2
